:truck: Chore: custom response 파일 수정 및 분리
